# services/gateway/docs/swagger.yaml
components:
  schemas:
...
    dto.CreateLocationRequest:
      properties:
        city:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        stateProvince:
          type: string
        venueName:
          type: string
        zones:
          items:
            $ref: '#/components/schemas/dto.CreateLocationZoneRequest'
          type: array
          uniqueItems: false
      required:
      - city
      - country
      - latitude
      - longitude
      - stateProvince
      - venueName
      - zones
      type: object
    dto.CreateLocationResponse:
      properties:
        id:
          type: string
      type: object
    dto.CreateLocationZoneRequest:
      properties:
        capacity:
          type: integer
        numberOfRows:
          type: integer
        seatsPerRow:
          type: integer
        zoneName:
          type: string
        zoneNumber:
          type: integer
      required:
      - capacity
      - numberOfRows
      - seatsPerRow
      - zoneName
      - zoneNumber
      type: object
...
    dto.HttpError:
      properties:
        error:
          type: string
      required:
      - error
      type: object
...
    dto.HttpResponse-dto_CreateLocationResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.CreateLocationResponse'
      required:
      - result
      type: object
...
    dto.HttpResponse-dto_ListLocationsResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.ListLocationsResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_LocationResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.LocationResponse'
      required:
      - result
      type: object
...
    dto.HttpResponse-dto_UpdateLocationResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.UpdateLocationResponse'
      required:
      - result
      type: object
    dto.ListLocationsResponse:
      properties:
        list:
          items:
            $ref: '#/components/schemas/dto.LocationResponse'
          type: array
          uniqueItems: false
      type: object
    dto.LocationResponse:
      properties:
        city:
          type: string
        country:
          type: string
        id:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        stateProvince:
          type: string
        venueName:
          type: string
        zones:
          items:
            $ref: '#/components/schemas/dto.ZoneResponse'
          type: array
          uniqueItems: false
      required:
      - city
      - country
      - id
      - latitude
      - longitude
      - stateProvince
      - venueName
      - zones
      type: object
...
    dto.UpdateLocationRequest:
      properties:
        city:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        stateProvince:
          type: string
        venueName:
          type: string
        zones:
          items:
            $ref: '#/components/schemas/dto.UpdateLocationZoneRequest'
          type: array
          uniqueItems: false
      required:
      - city
      - country
      - latitude
      - longitude
      - stateProvince
      - venueName
      - zones
      type: object
    dto.UpdateLocationResponse:
      properties:
        id:
          type: string
      type: object
    dto.UpdateLocationZoneRequest:
      properties:
        capacity:
          type: integer
        numberOfRows:
          type: integer
        seatsPerRow:
          type: integer
        zoneName:
          type: string
        zoneNumber:
          type: integer
      required:
      - capacity
      - numberOfRows
      - seatsPerRow
      - zoneName
      - zoneNumber
      type: object
...
    dto.ZoneResponse:
      properties:
        capacity:
          type: integer
        numberOfRows:
          type: integer
        seatsPerRow:
          type: integer
        zoneName:
          type: string
        zoneNumber:
          type: integer
      required:
      - capacity
      - numberOfRows
      - seatsPerRow
      - zoneName
      - zoneNumber
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: A Concert Gateway API Documentation
  title: A Concert Gateway
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/locations:
    get:
      description: List Locations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_ListLocationsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: List Locations
      tags:
      - locations
    post:
      description: Create Location
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.CreateLocationRequest'
        description: Create location request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_CreateLocationResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create Location
      tags:
      - locations
  /v1/locations/{id}:
    delete:
      description: Delete Location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Delete Location
      tags:
      - locations
    get:
      description: Get Location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_LocationResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: Get Location
      tags:
      - locations
    put:
      description: Update Location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.UpdateLocationRequest'
        description: Update location request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_UpdateLocationResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Update Location
      tags:
      - locations